fix(api): entity endpoints accept identifier form + int64 id consistency (#174)
* fix(api): accept entity identifier form in URLs + int64 id consistency (#172, #173)

Task/incident/corrective-action/audit-finding/legal endpoints parsed :id with raw
strconv.Atoi, so only a numeric id worked and the identifier form (TASK-6, …) 400'd.
Route them through parseID (extended to strip every NextIdentifier prefix incl.
legacy AST-), and standardize those entities' id to int64 to match suppliers/
assets/risks/systems — so parseID drops in and the whole API is on one id
vocabulary.

Also fixes the AST-/ASSET- split: the suggestion-apply path now uses the shared
nextIdentifierTx(…, "asset") helper (ASSET-) instead of a hardcoded AST-.

Regression test asserts identifier and numeric URLs resolve the same record for
tasks/incidents/CAs/legal (get/put/delete). go build + go unit tests green.

Closes #172. Closes #173.

* fix(api): resolve seq-based identifiers via lookup, not prefix-strip (#174)

The identifier-URL tests failed: parseID stripped the prefix and used the numeric
SUFFIX as the primary key, but for task/incident/corrective-action/legal the
suffix is a per-org identifier_sequences value, not the id — they diverge (e.g.
after rolled-back inserts), so GET /tasks/TASK-8 looked up id=8 and 404'd.

Adds resolveTaskID/resolveIncidentID/resolveCorrectiveActionID/resolveLegalID
that pass a numeric param straight through and look up the PREFIX-<seq> form via
GetXByIdentifier (added GetTaskByIdentifier; the others already existed). Findings
keep parseID — FIND-<id> uses the id itself, so the suffix IS the key. parseID
trimmed back to the prefixes whose suffix equals the id (FIND- + the pre-existing
supplier/asset/risk/system set).

Note: the same latent bug exists in the pre-existing supplier/asset/risk/system
parseID handlers (seq != id) — out of scope here, separate issue.

* fix(api): address #174 review — 404 vs 400, suggestion-apply lookup, comment

- resolve*ID: add errInvalidID sentinel + prefix check. A malformed param (no
  digits, wrong/no prefix) → 400; a well-formed-but-missing identifier returns the
  lookup error, which call sites now map to 404 — matching the numeric path.
- suggestion-apply: applyIncidentUpdate/applyCorrActiveUpdate/applyTaskUpdate
  resolved sg.EntityID via parseEntityID (suffix-as-id) — the same seq!=id bug.
  Now resolve via GetIncidentByIdentifier/GetCorrectiveActionByIdentifier/
  GetTaskByIdentifier (all exist).
- soften the parseID comment: supplier/asset/risk/system happen to be seq==id, not
  a structural guarantee.

* fix(api): suggestion-apply resolves entity_id in both id + identifier form (#174)

My earlier review commit switched applyIncidentUpdate/applyCorrActiveUpdate/
applyTaskUpdate to GetXByIdentifier(sg.EntityID). But sg.EntityID isn't always the
identifier form — test_corrective_writepath / test_incident_writepath post
entity_id: str(ca_id) (the numeric id), and GetByIdentifier("7") looks up
WHERE identifier='7' → no rows → the 'corrective action 7 not found' 500 in CI.

Route through the resolve*ID helpers instead, which accept EITHER a numeric id OR
the PREFIX-<seq> identifier (numeric passes straight to Get*, identifier resolves
via GetXByIdentifier). Fixes both the numeric-entity_id tests and the SPA's
identifier-form suggestions.

+++ internal/isms/api/api_references.go
@@ -294,7 +294,7 @@ func (s *Server) resolveEntityTitle(ctx context.Context, orgID int, entityType, 
 		if err != nil {
 			return entityID
 		}
-		f, err := s.db.GetAuditFinding(ctx, orgID, id)
+		f, err := s.db.GetAuditFinding(ctx, orgID, int64(id))
 		if err != nil {
 			return entityID
 		}
@@ -316,7 +316,7 @@ func (s *Server) resolveEntityTitle(ctx context.Context, orgID int, entityType, 
 		if err != nil {
 			return entityID
 		}
-		t, err := s.db.GetTask(ctx, orgID, id)
+		t, err := s.db.GetTask(ctx, orgID, int64(id))
 		if err != nil {
 			return entityID
 		}
